refactor(client): create MusicPlaylistEditorRoomSearchResult and use it in each mpe search engine
